refactor: extract tree info list component
