+++ pages/trees/[id].tsx
@@ -1,11 +1,10 @@
-import { FC, ReactElement, ReactNode } from 'react'
+import { ReactElement, ReactNode } from 'react'
 import { MapLayout } from '@layouts/MapLayout'
 import classNames from 'classnames'
 import { GetServerSideProps, NextPage } from 'next'
 import { TreeDataType } from '@lib/requests/getTreeData'
 import { useNowcastData } from '@lib/hooks/useNowcastData'
 import { TreeInfoHeader } from '@components/TreeInfoHeader'
-import { DataListItem } from '@components/DataListItem'
 import { mapSuctionTensionToStatus } from '@lib/utils/mapSuctionTensionToStatus'
 import { GroundLayersViz } from '@components/GroundLayersViz'
 import { useRouter } from 'next/router'
@@ -22,22 +21,13 @@ import csrf from '@lib/api/csrf'
 import { useForecastData } from '@lib/hooks/useForecastData'
 import { combineNowAndForecastData } from '@lib/utils/forecastUtil/forecastUtil'
 import { useTreeRainAmount } from '@lib/hooks/useTreeRainAmount'
-import { TreeRainAmountType } from '@lib/requests/getTreeRainAmount'
-import { MappedNowcastRowsType } from '@lib/utils/mapRowsToDepths'
 import { useTreeData } from '@lib/hooks/useTreeData'
 import { mapRawQueryToState } from '@lib/utils/queryUtil'
 import { Head } from '@components/Head'
-import { DatavisIcon } from '@components/DatavisIcons/DatavisIcon'
-import { normalizeValue } from '@lib/utils/normalizeValue'
 import { ParkTreeHint } from '@components/ParkTreeHint'
-import {
-  useShadingData,
-  useShadingDataReturnType,
-} from '@lib/hooks/useShadingData'
-import {
-  useWateringData,
-  useWateringDataReturnType,
-} from '@lib/hooks/useWateringData'
+import { useShadingData } from '@lib/hooks/useShadingData'
+import { useWateringData } from '@lib/hooks/useWateringData'
+import { TreeInfoList } from '@components/TreeInfoList'
 
 interface TreePageComponentPropType {
   treeId: TreeDataType['id']
@@ -83,138 +73,6 @@ export const getServerSideProps: GetServerSideProps<
       notFound: true,
     }
   }
-}
-
-const NoDataIndicator: FC = () => <span>–</span>
-
-const InfoList: FC<{
-  treeData: TreeDataType | null
-  treeDataIsLoading: boolean
-  treeDataError: Error | null
-  nowcastData: MappedNowcastRowsType | null
-  nowcastIsLoading: boolean
-  nowcastError: Error | null
-  rainData: TreeRainAmountType | null
-  rainIsLoading: boolean
-  rainError: Error | null
-  shadingData: useShadingDataReturnType['data']
-  shadingIsLoading: useShadingDataReturnType['isLoading']
-  shadingError: useShadingDataReturnType['error']
-  wateringData: useWateringDataReturnType['data']
-  wateringDataIsLoading: useWateringDataReturnType['isLoading']
-  wateringDataError: useWateringDataReturnType['error']
-}> = ({
-  treeData,
-  treeDataIsLoading,
-  rainData,
-  rainIsLoading,
-  shadingData,
-  shadingIsLoading,
-  wateringData,
-  wateringDataIsLoading,
-}) => {
-  const { t } = useTranslation('common')
-
-  // Original shadingData is between 0 - 1, so we multiply by 100:
-  const shadingValue = !shadingIsLoading && shadingData && shadingData * 100
-  const rainValue = !rainIsLoading && rainData
-  const treeDiscValue = !treeDataIsLoading && treeData?.baumscheibe
-  const trunkCircumferenceValue = !treeDataIsLoading && treeData?.stammumfg
-  const wateringValue = !wateringDataIsLoading && wateringData
-
-  return (
-    <ul className="relative z-10 bg-white">
-      <DataListItem
-        title={t(`treeView.infoList.shading.label`)}
-        subtitle={t(`treeView.infoList.shading.hint`)}
-        datavisIcon={
-          shadingValue ? (
-            <DatavisIcon
-              iconType="clock"
-              iconValue={normalizeValue(shadingValue, [0, 100])}
-              valueLabel={t(`treeView.infoList.shading.value`, {
-                value: shadingValue.toFixed(0),
-              })}
-            />
-          ) : (
-            <NoDataIndicator />
-          )
-        }
-      />
-      <DataListItem
-        title={t(`treeView.infoList.rainAmount.label`)}
-        subtitle={t(`treeView.infoList.rainAmount.hint`)}
-        datavisIcon={
-          rainValue ? (
-            <DatavisIcon
-              iconType="water-drops"
-              iconValue={Math.round(
-                normalizeValue(rainValue, [0, 500], [0, 5])
-              )}
-              valueLabel={t(`treeView.infoList.rainAmount.value`, {
-                value: rainValue.toFixed(1),
-              })}
-            />
-          ) : (
-            <NoDataIndicator />
-          )
-        }
-      />
-      <DataListItem
-        title={t(`treeView.infoList.wateringAmount.label`)}
-        subtitle={t(`treeView.infoList.wateringAmount.hint`)}
-        datavisIcon={
-          wateringValue ? (
-            <DatavisIcon
-              iconType="water-drops"
-              iconValue={Math.round(
-                normalizeValue(wateringValue, [0, 500], [0, 5])
-              )}
-              valueLabel={t(`treeView.infoList.wateringAmount.value`, {
-                value: wateringValue.toFixed(),
-              })}
-            />
-          ) : (
-            <NoDataIndicator />
-          )
-        }
-      />
-      <DataListItem
-        title={t(`treeView.infoList.treeDisc.label`)}
-        subtitle={t(`treeView.infoList.treeDisc.hint`)}
-        datavisIcon={
-          treeDiscValue ? (
-            <DatavisIcon
-              iconType="square"
-              iconValue={normalizeValue(treeDiscValue, [0, 10])}
-              valueLabel={t(`treeView.infoList.treeDisc.value`, {
-                value: treeData?.baumscheibe,
-              })}
-            />
-          ) : (
-            <NoDataIndicator />
-          )
-        }
-      />
-      <DataListItem
-        title={t(`treeView.infoList.trunkCircumference.label`)}
-        subtitle={t(`treeView.infoList.trunkCircumference.hint`)}
-        datavisIcon={
-          trunkCircumferenceValue ? (
-            <DatavisIcon
-              iconType="circle"
-              iconValue={normalizeValue(trunkCircumferenceValue, [0, 800])}
-              valueLabel={t(`treeView.infoList.trunkCircumference.value`, {
-                value: trunkCircumferenceValue,
-              })}
-            />
-          ) : (
-            <NoDataIndicator />
-          )
-        }
-      />
-    </ul>
-  )
 }
 
 const TreePage: TreePageWithLayout = ({ treeId, csrfToken }) => {
@@ -399,7 +257,7 @@ const TreePage: TreePageWithLayout = ({ treeId, csrfToken }) => {
               {
                 name: t('treeView.tabs.0'),
                 content: (
-                  <InfoList
+                  <TreeInfoList
                     treeData={treeData}
                     treeDataIsLoading={treeDataLoading}
                     treeDataError={treeDataError}
